Transferred oid fetching from ApplicationsServlet to Oppijanumerorekisteri.
