+++ src/main/scala/fi/vm/sade/hakemuseditori/oppijanumerorekisteri/oppijanumerorekisteri.scala
@@ -2,12 +2,16 @@ package fi.vm.sade.hakemuseditori.oppijanumerorekisteri
 
 import java.util.concurrent.TimeUnit
 
+import fi.vm.sade.hakemuseditori.json.JsonFormats
 import fi.vm.sade.omatsivut.OphUrlProperties
 import fi.vm.sade.omatsivut.config.AppConfig.AppConfig
 import fi.vm.sade.utils.cas.{CasAuthenticatingClient, CasClient, CasParams}
+import fi.vm.sade.utils.slf4j.Logging
 import org.http4s.Method.GET
 import org.http4s.client.blaze
-import org.http4s.{Request, Uri}
+import org.http4s.{Header, Headers, Request, Uri}
+import org.json4s
+import org.json4s._
 import org.json4s.JsonAST.{JNull, JObject, JString, JValue}
 import org.json4s.jackson.JsonMethods
 import org.json4s.{DefaultFormats, Reader, Writer}
@@ -39,6 +43,7 @@ object Henkilo {
 
 trait OppijanumerorekisteriService {
   def henkilo(personOid: String): Henkilo
+  def fetchAllOids(pOid: String): List[String]
 }
 
 trait OppijanumerorekisteriComponent {
@@ -46,9 +51,11 @@ trait OppijanumerorekisteriComponent {
 
   class StubbedOppijanumerorekisteriService extends OppijanumerorekisteriService {
     override def henkilo(personOid: String): Henkilo = ???
+    override def fetchAllOids(pOid: String): List[String] = List("1.2.246.562.24.14229104472")
   }
 
-  class RemoteOppijanumerorekisteriService(config: AppConfig) extends OppijanumerorekisteriService {
+  class RemoteOppijanumerorekisteriService(config: AppConfig) extends OppijanumerorekisteriService with JsonFormats with Logging {
+    import org.json4s.jackson.JsonMethods._
     private val blazeHttpClient = blaze.defaultClient
     private val casClient = new CasClient(config.settings.securitySettings.casUrl, blazeHttpClient)
     private val casParams = CasParams(
@@ -63,6 +70,7 @@ trait OppijanumerorekisteriComponent {
       Some("omatsivut.omatsivut.backend"),
       "JSESSIONID"
     )
+    private val callerIdHeader = Header("Caller-Id", "omatsivut.omatsivut.backend")
     implicit private val formats = DefaultFormats
     implicit private val henkiloReader = Henkilo.henkiloReader
 
@@ -74,6 +82,51 @@ trait OppijanumerorekisteriComponent {
             case r => Task.fail(new RuntimeException(s"Failed to get henkilö for $personOid: ${r.toString()}"))
           }
         }).attemptRunFor(Duration(10, TimeUnit.SECONDS)).fold(throw _, x => x)
+    }
+
+    private def uriFromString(url: String): Uri = {
+      Uri.fromString(url).toOption.get
+    }
+    private def runHttp[ResultType](request: Request)(decoder: (Int, String, Request) => ResultType): Task[ResultType] = {
+      httpClient.fetch(request)(r => r.as[String].map(body => decoder(r.status.code, body, request)))
+    }
+
+    override def fetchAllOids(pOid: String): List[String] = {
+        implicit val formats = DefaultFormats
+        val timeout = 1000*30L
+
+        val masterRequest: Request = Request(
+          uri = uriFromString(OphUrlProperties.url("oppijanumerorekisteri-service.henkilo-master", pOid)),
+          headers = Headers(callerIdHeader))
+
+        val masterOid: String = runHttp[Option[String]](masterRequest) {
+          case (200, resultString, _) =>
+            val f: json4s.JValue = parse(resultString).asInstanceOf[JObject]
+            val oid = f \ "oidHenkilo"
+            Some(oid.extract[String])
+          case (code, responseString, _) =>
+            logger.error("Failed to fetch master oid for user oid {}, response was {}, {}", pOid, Integer.toString(code), responseString)
+            None
+        }.runFor(timeoutInMillis = timeout).getOrElse(pOid)
+
+        val slaveRequest: Request = Request(
+          uri = uriFromString(OphUrlProperties.url("oppijanumerorekisteri-service.henkilo-slaves", masterOid)),
+          headers = Headers(callerIdHeader))
+
+        val allOids: List[String] = runHttp(slaveRequest) {
+          case (200, resultString, _) =>
+            val slaveOids: Seq[String] = parse(resultString).extract[List[JObject]]
+              .map(obj => {
+                val oidObj = obj \ "oidHenkilo"
+                oidObj.extract[String]
+              })
+            List(masterOid) ++ slaveOids
+          case (code, responseString, _) =>
+            logger.error("Failed to fetch slave OIDs for user oid {}, response was {}, {}", masterOid, Integer.toString(code), responseString)
+            List(masterOid)
+        }.runFor(timeoutInMillis = timeout)
+
+      return allOids
     }
   }
 }
